refactor: rename and align classnames
