+++ src/apps/search-result/search-result.mount.ts
@@ -1,5 +1,5 @@
 import addMount from "../../core/addMount";
-import SearchResultV2 from "../search-result/SearchResult.entry";
+import SearchResult from "../search-result/SearchResult.entry";
 
 // Using V2 component with sidebar facets
-addMount({ appName: "search-result", app: SearchResultV2 });
+addMount({ appName: "search-result", app: SearchResult });
